Consolidate Robot tests
